+++ CHANGELOG.md
@@ -7,19 +7,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Audio Management
 - Camera Controls (Rotate on Input Axis, Translate on Input)
 - CDI Logo
+- Chance component for randomizing events
 - CHANGELOG.md, LICENSE.md, README.md, Third Party Notices.md, and package.json
 - Controller Models, Materials, and Prefabs
 - Debug Controls
 - Editor Config
 - Fade Material Component
 - FPS Canvas and Counter
+- Handle Start component
 - Indexed Collections (GameObject, PlayableDirector, UnityEvent)
 - Input Prefabs (Event System, Gaze Input Module, XR Rig)
+- Lock Transform component
 - Math Constants and Extensions
 - Particle System Controller and Extensions
 - Sample Materials and Textures (Grid, Skybox)
+- Spawn Object component
 - Time Scaler
 
 [unreleased]: https://github.com/CDInstitute/CDIToolkit/compare/53afd0277bbaf5d926214c861a0445fb7f8b68a3...HEAD